feat request에 사용하는 함수들 파일로 분리
